cms-admin: rename ChooseFile(s)Dialog to ChooseDamFile(s)Dialog
Renames the DAM file picker dialogs to make their scope explicit:
ChooseFileDialog → ChooseDamFileDialog,
ChooseFilesDialog → ChooseDamFilesDialog,
BaseChooseFromDamDialog → BaseChooseDamFileDialog. The old
ChooseFileDialog name is kept as a deprecated re-export.
