+++ .changeset/export-choose-dam-files-dialog.md
@@ -2,12 +2,12 @@
 "@comet/cms-admin": minor
 ---
 
-Export `ChooseFilesDialog`
+Export `ChooseDamFilesDialog`
 
 Allows building custom multi-file picker UIs on top of the DAM file dialog (e.g. bulk-adding files to a list block). Pair with `damMultiFileFieldFragment` and a Dam file query to fetch the picked file metadata.
 
 ```tsx
-import { ChooseFilesDialog } from "@comet/cms-admin";
+import { ChooseDamFilesDialog } from "@comet/cms-admin";
 
-<ChooseFilesDialog open={open} onClose={onClose} onConfirm={(fileIds) => ...} initialFileIds={[]} allowedMimetypes={["image/jpeg"]} />
+<ChooseDamFilesDialog open={open} onClose={onClose} onConfirm={(fileIds) => ...} initialFileIds={[]} allowedMimetypes={["image/jpeg"]} />
 ```
